style: estilizacao da tela de grafico iniciada
