+++ Regula.sun - front/src/router/index.js
@@ -2,7 +2,7 @@ import { createRouter, createWebHistory } from "vue-router";
 import HomeView from "../views/HomeView.vue";
 import LoginView from "../views/LoginView.vue";
 import AdminView from "../views/AdminView.vue";
-import Grafico from "../views/GraficoView.vue";
+import GraficoView from "../views/GraficoView.vue";
 // import Boletim from "../views/BoletimView.vue";
 import { validarTokenAcesso } from "../service/autenticacao";
 import store from "../store/index.js";
@@ -18,7 +18,7 @@ const router = createRouter({
         {
             path: "/grafico",
             name: "grafico",
-            component: Grafico,
+            component: GraficoView,
         },
         // {
         //   path: "/boletim",
